魚修正中

+++ bf2-4e/source/Fish.h
@@ -6,8 +6,8 @@ private:
 	float fish_x;
 	float fish_y;
 
-	int fish_image[10];
-	float fish_animation_count;//フレームレートカウント
+	int fish_image [10] ;
+	int fish_animation_count;//フレームレートカウント
 	int fish_flg;		//0..右移動 1..左移動 2..浮上 3..プレイヤー喰らう
 
 	int now_image;							// 今から表示される画像
